+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to `EchoQuery` will be documented in this file.
 ## Last update 07-04-2024
 
 ### Added
-- Where statement less than (<>) comparison operator implemented
+- Where statement comparison operator refactor
 
 ### Deprecated
 - Nothing